+++ projects/angular-sdk/src/apis/hopper-cloud-airline/v1/api/customer.service.ts
@@ -44,7 +44,7 @@ import { Configuration }                                     from '../configurat
 @Injectable()
 export class CustomerService {
 
-    protected basePath = '/airline/v1.0';
+    protected basePath = '/airline/v1.1';
     public defaultHeaders = new HttpHeaders();
     public configuration = new Configuration();