+++ src/app/pages/guides/cfar-offer-banner/cfar-offer-banner.component.ts
@@ -104,7 +104,7 @@ export class CfarOfferBannerPageComponent extends CommonGuidesComponent {
       {
         name: 'itineraries',
         description: `
-          See <a target="_blank" aria-describedby="api documentation" href="https://airlines-api.staging.hopper.com/airline/v1.0/docs/index.html#operation/postCfar_offers">API documentation</a>
+          See <a target="_blank" aria-describedby="api documentation" href="https://airlines-api.staging.hopper.com/airline/v1.1/docs/index.html#operation/postCfar_offers">API documentation</a>
         `,
         required: true
       }